rename to agent helper, add comment, remove fanout and unsed func in aggregator
