+++ config/types_test.go
@@ -5,16 +5,16 @@ import (
 	"testing"
 	"time"
 
+	"github.com/influxdata/telegraf/agenthelper"
 	"github.com/influxdata/telegraf/config"
 	"github.com/influxdata/telegraf/models"
 	"github.com/influxdata/telegraf/plugins/processors/reverse_dns"
-	"github.com/influxdata/telegraf/testhelper"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConfigDuration(t *testing.T) {
 	c := config.NewConfig()
-	c.SetAgent(&testhelper.TestAgentController{})
+	c.SetAgent(&agenthelper.TestAgentController{})
 	err := c.LoadConfigData(context.Background(), context.Background(), []byte(`
 [[processors.reverse_dns]]
   cache_ttl = "3h"
