--- internal/channel/fanout.go
@@ -1,1 +1,0 @@
-package channel
