--- internal/channel/fanout_test.go
@@ -1,1 +1,0 @@
-package channel
