+++ plugins/parsers/json_v2/parser_test.go
@@ -9,11 +9,11 @@ import (
 	"testing"
 
 	"github.com/influxdata/telegraf"
+	"github.com/influxdata/telegraf/agenthelper"
 	"github.com/influxdata/telegraf/config"
 	"github.com/influxdata/telegraf/plugins/inputs"
 	"github.com/influxdata/telegraf/plugins/inputs/file"
 	"github.com/influxdata/telegraf/plugins/parsers/influx"
-	"github.com/influxdata/telegraf/testhelper"
 	"github.com/influxdata/telegraf/testutil"
 	"github.com/stretchr/testify/require"
 )
@@ -98,7 +98,7 @@ func TestData(t *testing.T) {
 				return &file.File{}
 			})
 			cfg := config.NewConfig()
-			cfg.SetAgent(&testhelper.TestAgentController{})
+			cfg.SetAgent(&agenthelper.TestAgentController{})
 			err = cfg.LoadConfigData(context.Background(), context.Background(), buf)
 			require.NoError(t, err)
